chore: simplify query imports
