+++ apps/passport-server/src/database/queries/telegram/delete.ts
@@ -46,3 +46,22 @@ export async function deleteTelegramChatTopic(
     [telegramChatId, telegramTopicId]
   );
 }
+
+/**
+ * Delete a verification.
+ */
+export async function deleteTelegramVerification(
+  client: Pool,
+  telegramUserId: number,
+  telegramChatId: number
+): Promise<void> {
+  await sqlQuery(
+    client,
+    `\
+    delete from telegram_bot_conversations
+    where telegram_user_id = $1
+    and telegram_chat_id = $2 
+    `,
+    [telegramUserId, telegramChatId]
+  );
+}
